requirements.txt
absl-py==1.4.0
anyio==3.6.2
astunparse==1.6.3
blis==0.7.9
cachelib==0.10.2
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
fastapi==0.93.0
Flask==2.2.3
Flask-Session==0.4.0
flatbuffers==23.3.3
fst-pso==1.8.1
grpcio==1.51.3
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
tensorflow
uvicorn==0.20.0
websockets==10.4
Pillow
pydantic
opencv-python
numpy
matplotlib
scikit-learn
